rename: swagger 관련 클래스 이동
